move function to the module
